move mainFrame exiting service to mainFrame, better logging
